Use `AK.Recertify` to recertify a `Key`
